+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Fury OAS3 Parser Changelog
 
+## Master
+
+### Enhancements
+
+- Support for "default" response status codes.
+
 ## 0.10.2 (2020-03-16)
 
 ### Enhancements